requirements.txt
flask
transformers
torch
python-dotenv
streamlit
requests
pandas
gunicorn